# remove max_block_size_ in train states (fix #3570)