This is to drop the unused max_block_size_ in train share states, and to fix issue #3570. The bug occurs when the number of training data is quite small.