@guolinke 2020-11-22
@shiyu1994 did you apply the fix we discussed last week?