@shiyu1994 2020-11-23
Restrict the `min_block_size_` to be `>= 32`. In comparison with #3522, it shows slightly better results.

| Dataset | This PR | Old Master (Before #3522) | Speedup |
| --- | --- | --- | --- |
|higgs|**136.13**±0.71 | 155.33±2.60 |  1.14 |
|yahoo|**115.28**±1.65 | 146.13±1.99 |  1.27 |
|msltr|**148.03**±1.19 | 163.52±1.29 |  1.10 |
|dataexpo_onehot|**70.58**±0.86 | 78.62±1.47 |  1.11 |
|allstate|**170.96**±1.86 | 182.61±2.44 |  1.07 |
|adult|**0.99**±0.02 | 1.17±0.04 |  1.18 |
|amazon|0.71±0.01 | 0.71±0.04 |  1.00 |
|appetency|**1.57**±0.07 | 1.62±0.11 |  1.03 |
|click|**18.29**±0.19 | 19.24±0.68 |  1.05 |
|internet|**0.33**±0.01 | 0.43±0.02 |  1.31 |
|kick|**3.73**±0.10 | 3.78±0.09 |  1.02 |
|upselling|3.28±0.14 | **3.11**±0.20 |  0.95 |
|nips_b|**57.59**±0.58 | 62.93±0.66 |  1.09 |
|nips_c|**8.36**±0.09 | 8.71±0.30 |  1.04 |
|year|**38.46**±0.62 | 46.94±0.61 |  1.22 |